backend/requirements.txt
# backend/requirements.txt - COMPLETE
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
boto3==1.34.0
google-generativeai==0.3.2
pandas==2.1.4
python-levenshtein==0.23.0
stripe==8.4.0
websockets==12.0